+++ Changelog.md
@@ -5,7 +5,7 @@
 [![Web Site](https://img.shields.io/badge/Markdown_Monster-WebSite-blue.svg)](https://markdownmonster.west-wind.com)
 
 ### 1.18.15
-*<small>not released yet</small>*
+*<small>September 3rd, 2019</small>*
 
 * **Pre-Configured Window Size Selection**  
 The Control Box now has a dropdown that lets you select a pre-configured Window size and resize your window to one of these sizes. The list is customizable and you can modify or add your own sizes. Resizing is smart enough to create windows that fit onto the active screen and translates for high DPI modes.
@@ -14,7 +14,8 @@ The Control Box now has a dropdown that lets you select a pre-configured Window 
 You can now backup your Markdown Monster configuration to Zip file or to a disk folder. This feature writes out all files from the configuration folder into the zip or folder as a backup mechanism.
 
 * **Add Open Document File Change Monitoring**  
-The open document now actively tracks changes that are made externally and updates the document if the document has no changes. If the doc has changes you'll still be prompted when the editor is activated to the effect that the document has been changed externally.
+The open document now actively tracks changes that are made externally and updates the document if the document has no changes, now even if the editor is not activated so you can see changes made in another application in real time. As before, if the document has changes updates are held off until you activate the document at which point you get a prompt on whether to reload, or keep the editor document.
+
 
 ### 1.18.14
 *<small>August 27th, 2019</small>*